requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
